+++ lib/tenstorrent/bh_arc/flash_info_table.c
@@ -15,7 +15,7 @@ static FlashInfoTable flash_info_table;
 int load_flash_info_table(uint8_t *buffer_space, uint32_t buffer_size)
 {
 	static const char flashInfoTableTag[TT_BOOT_FS_IMAGE_TAG_SIZE] = "flshinfo";
-	uint32_t bin_size = 0;
+	size_t bin_size = 0;
 
 	if (tt_boot_fs_get_file(&boot_fs_data, flashInfoTableTag, buffer_space, buffer_size,
 				&bin_size) != TT_BOOT_FS_OK) {
