+++ lib/tenstorrent/bh_arc/read_only_table.c
@@ -25,7 +25,7 @@ static ReadOnly read_only_table;
 /* Loader function that deserializes the fw table bin from the SPI filesystem */
 int load_read_only_table(uint8_t *buffer_space, uint32_t buffer_size)
 {
-	uint32_t bin_size = 0;
+	size_t bin_size = 0;
 	static const char readOnlyTag[TT_BOOT_FS_IMAGE_TAG_SIZE] = "boardcfg";
 
 	if (tt_boot_fs_get_file(&boot_fs_data, readOnlyTag, buffer_space, buffer_size, &bin_size) !=
